Move redux utils to ducks and create app services to handle user login
