+++ src/ducks/utils.js
@@ -1,0 +1,3 @@
+export const createActionTypes = actionTypes => actionTypes
+	.map(type => ({ [type]: `module/${type}`	}))
+	.reduce((types, type) => ({ ...types, ...type }), {});
